requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
anthropic==0.42.0
httpx==0.28.1
pillow==11.1.0
python-dotenv==1.0.1
python-multipart==0.0.19
jinja2==3.1.5
pydantic==2.10.4
